factor out utils/ validated crbi regressor

+++ pnc/data_saver.py
@@ -19,13 +19,14 @@ class DataSaver(metaclass=MetaSingleton):
     Data Saver:
         add topics --> advance
     """
-    def __init__(self):
+    def __init__(self, filename='pnc.pkl'):
         self._history = dict()
         if not os.path.exists('data'):
             os.makedirs('data')
         for f in os.listdir('data'):
-            os.remove('data/' + f)
-        self._file = open('data/history.pkl', 'ab')
+            if f == filename:
+                os.remove('data/' + f)
+        self._file = open('data/' + filename, 'ab')
 
     def add(self, key, value):
         self._history[key] = value
